# optimizations around memory allocation and footprint