--- a/benchmarks/pom.xml
+++ b/benchmarks/pom.xml
@@ -72,6 +72,15 @@
           </execution>
         </executions>
       </plugin>
+
+      <!-- no need to deploy the benchmarks jar anywhere -->
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-deploy-plugin</artifactId>
+        <configuration>
+          <skip>true</skip>
+        </configuration>
+      </plugin>
     </plugins>
   </build>