--- a/core/src/main/java/com/spotify/missinglink/datamodel/ClassTypeDescriptor.java
+++ b/core/src/main/java/com/spotify/missinglink/datamodel/ClassTypeDescriptor.java
@@ -23,7 +23,7 @@ public class ClassTypeDescriptor implements TypeDescriptor {
 
   private final String className;
 
-  public ClassTypeDescriptor(String className) {
+  ClassTypeDescriptor(String className) {
     this.className = Preconditions.checkNotNull(className).replace('/', '.');
     if (className.endsWith(";")) {
       throw new InputMismatchException(