… core/src/test/java/com/spotify/missinglink/ArtifactLoaderTest.java
… core/src/test/java/com/spotify/missinglink/ConflictCheckerTest.java
… core/src/test/java/com/spotify/missinglink/Simple.java
… core/src/test/java/com/spotify/missinglink/TypeDescriptorTest.java
… maven-plugin/src/test/java/com/spotify/missinglink/maven/CheckMojoTest.java
… pom.xml