@pettermahlen 2015-06-20
Nice! I think it looks great. I wonder if we could gain further benefits from GC tuning - @spkrka's comment made me wonder if we could for instance gain something from ensuring we get the cheapest possible young generation collections - very high allocation pressure will lead to objects being copied between survivor spaces and probably also promoted too soon. This change should reduce the allocation pressure, but we could maybe do more to ensure that objects actually die before they get unnecessarily copied or promoted.