@mattnworb 2015-06-22
@pettermahlen I haven't looked into it but I am not sure that a maven plugin can influence the GC settings of Maven at all - other than instructing users on what `MAVEN_OPTS` settings to tweak.