@pettermahlen 2015-06-23
I was thinking it might be possible to control jvm  args in a forked execution. At least, the jetty plugin does that. Could be something to explore.